Fixed ReallocAndAppend buf handling and added overflow checks

+++ src/util.h
@@ -29,7 +29,7 @@ int IsPortInUse(struct PacketInfo *pi);
 char *ReportPacketType(const struct tcphdr *);
 char *ErrnoString(char *buf, const size_t buflen);
 int CreateDateTime(char *buf, const size_t size);
-__attribute__((format(printf, 3, 4))) char *ReallocAndAppend(char *filter, int *filterLen, const char *append, ...);
+__attribute__((format(printf, 3, 4))) char *ReallocAndAppend(char *filter, size_t *filterLen, const char *append, ...);
 
 #ifndef NDEBUG
 void DebugWritePacketToFs(const struct PacketInfo *pi);
